refactor(blorktools): repacking
