+++ packages/blorktools/src/asset_debugger/util/animation/state/animation-state.js
@@ -1,7 +1,7 @@
 import * as THREE from 'three';
-import { showStatus } from "../../modals/html-editor-modal/html-editor-modal";
-import { createMeshInfoPanel } from '../../modals/html-editor-modal/mesh-info-panel-util';
-import { createTextureFromIframe, injectUnifiedAnimationDetectionScript } from './html2canvas-util';
+import { showStatus } from "../../../modals/html-editor-modal/html-editor-modal";
+import { createMeshInfoPanel } from '../../../modals/html-editor-modal/mesh-info-panel-util';
+import { createTextureFromIframe, injectUnifiedAnimationDetectionScript } from '../../custom-animation/html2canvas-util';
 
 export const ANALYSIS_DURATION_MS = 30000; // 30 seconds - matches preRenderMaxDuration
